+++ controllers/moduleGroupsController.js
@@ -1,6 +1,8 @@
 const responseHelper = require('../helpers/responseHelper');
+const errorResponseHelper = require('../helpers/errorResponseHelper');
 const moduleGroupService = require('../services/moduleGroupService');
 const db = require('../database/database');
+const { updateModule } = require('../services/moduleService');
 
 exports.postModuleGroups = async (req, res, next) => {
   const transaction = await db.sequelize.transaction();
@@ -13,39 +15,54 @@ exports.postModuleGroups = async (req, res, next) => {
     return responseHelper(res, 201, 'Successfully created', createdModuleGroup);
   } catch (error) {
     transaction.rollback();
-    return next(error);
+    return errorResponseHelper(res, next, error);
   }
 };
 
 exports.putModuleGroups = async (req, res, next) => {
-  const transaction = await db.sequelize.transaction();
   const moduleGroup_id = req.query.moduleGroupId;
+  const transaction = await db.sequelize.transaction();
 
   try {
+    if (!moduleGroup_id) {
+      throw new Error('No module group given');
+    }
+
     const ModuleGroupToUpdate = req.body;
     const updatedModuleGroup = await moduleGroupService.updateModuleGroup(transaction, {
       ...ModuleGroupToUpdate,
       moduleGroup_id,
     });
+    if (!updatedModuleGroup) {
+      throw new Error('No module group to update');
+    }
 
     transaction.commit();
     return responseHelper(res, 200, 'Successfully updated', updatedModuleGroup);
   } catch (error) {
     transaction.rollback();
-    return next(error);
+    return errorResponseHelper(res, next, error);
   }
 };
 
 exports.deleteModuleGroups = async (req, res, next) => {
-  const transaction = await db.sequelize.transaction();
   const moduleGroupId = req.query.moduleGroupId;
+  const transaction = await db.sequelize.transaction();
 
   try {
+    if (!moduleGroupId) {
+      throw new Error('No module group given');
+    }
+
     const deletedModulecatalog = await moduleGroupService.deleteModuleGroup(transaction, moduleGroupId);
+    if (!deletedModulecatalog) {
+      throw new Error('No module group found to delete');
+    }
+
     transaction.commit();
     return responseHelper(res, 200, 'Successfully deleted', deletedModulecatalog);
   } catch (error) {
     transaction.rollback();
-    return next(error);
+    return errorResponseHelper(res, next, error);
   }
 };
