+++ controllers/semestersController.js
@@ -1,5 +1,6 @@
-const db = require('../database/database');
 const responseHelper = require('../helpers/responseHelper');
+const errorResponseHelper = require('../helpers/errorResponseHelper');
+const db = require('../database/database');
 const semesterService = require('../services/semesterService');
 const { checkSemesterEditAuthorization, checkCourseEditAuthorization } = require('../helpers/checkAuthorizationHelper');
 
@@ -15,12 +16,12 @@ exports.postSemesters = async (req, res, next) => {
     }
 
     const createdSemester = await semesterService.createSemester(transaction, SemesterToCreate);
-    transaction.commit();
 
+    transaction.commit();
     return responseHelper(res, 201, 'Successfully created', createdSemester);
   } catch (error) {
     transaction.rollback();
-    return next(error);
+    return errorResponseHelper(res, next, error);
   }
 };
 
@@ -30,7 +31,12 @@ exports.putSemesters = async (req, res, next) => {
   const transaction = await db.sequelize.transaction();
 
   try {
+    if (!semesterId) {
+      throw new Error('No semester given');
+    }
+
     const semesterToUpdate = req.body;
+
     if (
       !(await checkSemesterEditAuthorization(directorOfStudiesId, semesterId)) ||
       !(await checkCourseEditAuthorization(directorOfStudiesId, semesterToUpdate.course_id))
@@ -38,35 +44,47 @@ exports.putSemesters = async (req, res, next) => {
       transaction.rollback();
       return responseHelper(res, 403, 'You are not authorized to update this semester');
     }
+
     const updatedSemester = await semesterService.updateSemester(transaction, {
       ...semesterToUpdate,
       semester_id: semesterId,
     });
-    transaction.commit();
+    if (!updatedSemester) {
+      throw new Error('No semester found to update');
+    }
 
+    transaction.commit();
     return responseHelper(res, 200, 'Successfully updated', updatedSemester);
   } catch (error) {
     transaction.rollback();
-    return next(error);
+    return errorResponseHelper(res, next, error);
   }
 };
 
 exports.deleteSemesters = async (req, res, next) => {
   const semesterId = req.query.semesterId;
   const directorOfStudiesId = req.token.directorOfStudies_id;
-
   const transaction = await db.sequelize.transaction();
+
   try {
+    if (!semesterId) {
+      throw new Error('No semester given');
+    }
+
     if (!(await checkSemesterEditAuthorization(directorOfStudiesId, semesterId))) {
       transaction.rollback();
       return responseHelper(res, 403, 'You are not authorized to delete this semester');
     }
-    const deletedSemester = await semesterService.deleteSemester(transaction, semesterId);
-    transaction.commit();
 
+    const deletedSemester = await semesterService.deleteSemester(transaction, semesterId);
+    if (!deletedSemester) {
+      throw new Error('No semester found to delete');
+    }
+
+    transaction.commit();
     return responseHelper(res, 200, 'Successfully deleted', deletedSemester);
   } catch (error) {
     transaction.rollback();
-    return next(error);
+    return errorResponseHelper(res, next, error);
   }
 };
